+++ src/content-scripts/github.ts
@@ -16,7 +16,6 @@ import domUpdateWatch from "../utils/dom-utils/domUpdateWatcher";
 import injectDescriptionGeneratorButton from "../utils/dom-utils/addDescriptionGenerator";
 import injectChangeSuggestorButton from "../utils/dom-utils/changeSuggestorButton";
 import prEditWatch, { prReviewWatch } from "../utils/dom-utils/prWatcher";
-import { getAiDescription } from "./components/GenerateAIDescription/DescriptionGeneratorButton";
 
 const processGithubPage = async () => {
     if (prefersDarkMode(document.cookie)) {
@@ -56,27 +55,3 @@ const processGithubPage = async () => {
 };
 
 void processGithubPage();
-
-chrome.runtime.onMessage.addListener((msg, sender, sendResponse) => {
-    switch (msg.type) {
-        case "get_highlight": {
-            const title = (document.querySelector(".js-issue-title.markdown-title") as HTMLHeadingElement).innerText;
-
-            sendResponse(title);
-            break;
-        }
-        case "get_ai_description": {
-            const asyncRequest = async () => {
-                const aiText = await getAiDescription();
-
-                sendResponse(aiText);
-            };
-
-            asyncRequest().catch((e: Error | undefined) => {
-                sendResponse(e?.toString());
-                console.error(e);
-            });
-            return true;
-        }
-    }
-});